Try loading a wider variety of shared-object names
The "Python 3 compatible" code that I was using before is now, alas,
failing on Python 3.3 on my Ubuntu 13.04 laptop, because the new-style
shared object name, that I was looking for, is now back to being a
normal ".so" file.  So, with no way to guess what the right name is, I
am adding a loop over several possible names to try to select the right
one.